Changed driver.sh to make it easier to restart a run that had been terminated prematurely. Also improved time marks in driver output.

+++ standalone/examples/single_column_MARBL/BATS/driver.sh
@@ -2,6 +2,9 @@
 
 # ======================= SCRIPT PARAMETERS =======================
 
+# if 'true', then the script will erase old DART outputs and start clean
+START_CLEAN=false
+
 # MOM6 binary, keep this path absolute
 MOM6_BIN=~/work/cesm/cesm2_3_alpha12b+mom6_marbl/components/mom/standalone/build/intel-cheyenne/MOM6/MOM6
 
@@ -10,10 +13,10 @@ MOM6_BATS_DIR=$(pwd)                        # working directory for MOM6
 OBSSEQ_DIR=~/work/BATS_obsseq               # location of obs-sequence files
 
 # ensemble size
-ENS_SIZE=3
+ENS_SIZE=80
 
 # other
-LASTDAY_DART=147615 #147800     # last day of simulation (DART calendar)
+LASTDAY_DART=147977     # last day of simulation (DART calendar)
 MOM6_TO_DART=139157     # offset between MOM6 and DART calendars
 MOM6_TIMESTEP=3600.0    # timestep (seconds) to use when advancing ensemble members
 
@@ -21,64 +24,65 @@ MOM6_TIMESTEP=3600.0    # timestep (seconds) to use when advancing ensemble memb
 
 module load nco
 
-init_seconds=${SECONDS}
-
 echo ""
 echo "================================================================"
 echo "============================ DRIVER ============================"
 echo "================================================================"
 echo ""
 
-echo "backing up the initial ensemble..."
+if ${START_CLEAN}; then
+    echo "backing up the initial ensemble..."
 
-rm -rf ${MOM6_BATS_DIR}/ensemble_backup
-cp -r ${MOM6_BATS_DIR}/ensemble ensemble_backup
+    rm -rf ${MOM6_BATS_DIR}/ensemble_backup
+    cp -r ${MOM6_BATS_DIR}/ensemble ensemble_backup
 
-echo "deleting old DART output files..."
+    echo "deleting old DART output files..."
 
-rm -rf ${MOM6_BATS_DIR}/output/*
-rm -f ${MOM6_BATS_DIR}/DART/template_priorinf_mean.nc
-rm -f ${MOM6_BATS_DIR}/DART/template_priorinf_sd.nc
-rm -f ${MOM6_BATS_DIR}/DART/input_priorinf_mean.nc
-rm -f ${MOM6_BATS_DIR}/DART/input_priorinf_sd.nc
-rm -f ${MOM6_BATS_DIR}/DART/output_priorinf_mean.nc
-rm -f ${MOM6_BATS_DIR}/DART/output_priorinf_sd.nc
+    rm -rf ${MOM6_BATS_DIR}/output/*
+    mkdir ${MOM6_BATS_DIR}/output/member_0001_archive
+    rm -f ${MOM6_BATS_DIR}/DART/template_priorinf_mean.nc
+    rm -f ${MOM6_BATS_DIR}/DART/template_priorinf_sd.nc
+    rm -f ${MOM6_BATS_DIR}/DART/input_priorinf_mean.nc
+    rm -f ${MOM6_BATS_DIR}/DART/input_priorinf_sd.nc
+    rm -f ${MOM6_BATS_DIR}/DART/output_priorinf_mean.nc
+    rm -f ${MOM6_BATS_DIR}/DART/output_priorinf_sd.nc
 
-echo "configuring the DART namelist file..."
+    echo "configuring the DART namelist file..."
 
-sed -i "142 s%template_file = .*%template_file = '${MOM6_BATS_DIR}/ensemble/member_0001/RESTART/MOM.res.nc',%" ${MOM6_BATS_DIR}/DART/input.nml
-sed -i "32 s%input_state_files = .*%input_state_files = '',%" ${MOM6_BATS_DIR}/DART/input.nml
-sed -i "33 s%input_state_file_list = .*%input_state_file_list = '${MOM6_BATS_DIR}/DART/ensemble_members.txt',%" ${MOM6_BATS_DIR}/DART/input.nml
-sed -i "38 s%output_state_files = .*%output_state_files = '',%" ${MOM6_BATS_DIR}/DART/input.nml
-sed -i "39 s%output_state_file_list = .*%output_state_file_list = '${MOM6_BATS_DIR}/DART/ensemble_members.txt',%" ${MOM6_BATS_DIR}/DART/input.nml
-sed -i "47 s/ens_size = .*/ens_size = ${ENS_SIZE},/" ${MOM6_BATS_DIR}/DART/input.nml
-sed -i "42 s/num_output_state_members = .*/num_output_state_members = ${ENS_SIZE},/" ${MOM6_BATS_DIR}/DART/input.nml
-sed -i "58 s/num_output_obs_members = .*/num_output_obs_members = ${ENS_SIZE},/" ${MOM6_BATS_DIR}/DART/input.nml
-sed -i "49 s/perturb_from_single_instance = .*/perturb_from_single_instance = .false.,/" ${MOM6_BATS_DIR}/DART/input.nml
+    sed -i "142 s%template_file = .*%template_file = '${MOM6_BATS_DIR}/ensemble/member_0001/RESTART/MOM.res.nc',%" ${MOM6_BATS_DIR}/DART/input.nml
+    sed -i "32 s%input_state_files = .*%input_state_files = '',%" ${MOM6_BATS_DIR}/DART/input.nml
+    sed -i "33 s%input_state_file_list = .*%input_state_file_list = '${MOM6_BATS_DIR}/DART/ensemble_members.txt',%" ${MOM6_BATS_DIR}/DART/input.nml
+    sed -i "38 s%output_state_files = .*%output_state_files = '',%" ${MOM6_BATS_DIR}/DART/input.nml
+    sed -i "39 s%output_state_file_list = .*%output_state_file_list = '${MOM6_BATS_DIR}/DART/ensemble_members.txt',%" ${MOM6_BATS_DIR}/DART/input.nml
+    sed -i "47 s/ens_size = .*/ens_size = ${ENS_SIZE},/" ${MOM6_BATS_DIR}/DART/input.nml
+    sed -i "42 s/num_output_state_members = .*/num_output_state_members = ${ENS_SIZE},/" ${MOM6_BATS_DIR}/DART/input.nml
+    sed -i "58 s/num_output_obs_members = .*/num_output_obs_members = ${ENS_SIZE},/" ${MOM6_BATS_DIR}/DART/input.nml
+    sed -i "49 s/perturb_from_single_instance = .*/perturb_from_single_instance = .false.,/" ${MOM6_BATS_DIR}/DART/input.nml
 
-echo "generating inflation restart files..."
+    echo "generating inflation restart files..."
 
-back=$(pwd)
-cd ${MOM6_BATS_DIR}/DART
+    back=$(pwd)
+    cd ${MOM6_BATS_DIR}/DART
 
-echo ""
-echo "================================================================"
-echo "==================== FILL INFLATION RESTART ===================="
-echo "================================================================"
-echo ""
+    echo ""
+    echo "================================================================"
+    echo "==================== FILL INFLATION RESTART ===================="
+    echo "================================================================"
+    echo ""
 
-./fill_inflation_restart
+    ./fill_inflation_restart
 
-echo ""
-echo "================================================================"
-echo "============================ DRIVER ============================"
-echo "================================================================"
-echo ""
+    echo ""
+    echo "================================================================"
+    echo "============================ DRIVER ============================"
+    echo "================================================================"
+    echo ""
 
-cp input_priorinf_mean.nc template_priorinf_mean.nc
-cp input_priorinf_sd.nc template_priorinf_sd.nc
+    cp input_priorinf_mean.nc template_priorinf_mean.nc
+    cp input_priorinf_sd.nc template_priorinf_sd.nc
 
-cd ${back}
+    cd ${back}
+fi
 
 echo "setting MOM6 timestep to ${MOM6_TIMESTEP} seconds for each ensemble member..."
 
@@ -106,6 +110,9 @@ echo "beginning the assimilation loop..."
 
 while [ $currentday_dart -lt $LASTDAY_DART ]
 do
+    echo "archiving the state of member 1..."
+    cp ${MOM6_BATS_DIR}/ensemble/member_0001/RESTART/MOM.res.nc ${MOM6_BATS_DIR}/output/member_0001_archive/${currentday_dart}.res.nc
+
     echo "searching for an obs-sequence file for day ${currentday_mom6} (MOM6 calendar)..."
     echo "                                           ${currentday_dart} (DART calendar)..."
 
@@ -166,6 +173,8 @@ do
         sed -i "371 s/DAYMAX = .*/DAYMAX = ${tomorrow_mom6}/" ${MOM6_BATS_DIR}/ensemble/member_$(printf "%04d" ${i})/MOM_input
     done
 
+    init_seconds=${SECONDS}
+
     for i in $(seq ${ENS_SIZE}); do
         echo "advancing ensemble member ${i}..."
         cd ${MOM6_BATS_DIR}/ensemble/member_$(printf "%04d" ${i})
@@ -181,9 +190,7 @@ do
     current_seconds=${SECONDS}
     let diff=${current_seconds}-${init_seconds}
 
-    echo ""
-    echo "time since start: ${diff} seconds."
-    echo ""
+    echo "integration wall-time: ${diff} seconds."
 
     currentday_mom6=${tomorrow_mom6}
     currentday_dart=${tomorrow_dart}
